requirements.txt
transformers==4.30.2
cpm_kernels
torch>=2.0
# torchvision
# tensorboard
# cython
mdtex2html
sentencepiece
accelerate
paddlepaddle==2.4.2
paddleocr~=2.6.1.3
langchain==0.0.222
protobuf
unstructured[local-inference]
layoutparser[layoutmodels,tesseract]
nltk~=3.8.1
sentence-transformers
beautifulsoup4
icetk
faiss-cpu
gradio==3.35.2
gradio_client==0.2.7
fastapi
uvicorn~=0.21.1
pypinyin~=0.48.0
click~=8.1.3
tabulate
feedparser
azure-core
openai
peft
#bitsandbytes; platform_system != "Windows"
#llama-cpp-python==0.1.34; platform_system != "Windows"
#https://github.com/abetlen/llama-cpp-python/releases/download/v0.1.34/llama_cpp_python-0.1.34-cp310-cp310-win_amd64.whl; platform_system == "Windows"
pydantic~=1.10.7
starlette~=0.27.0
numpy~=1.23.5
tqdm~=4.65.0
requests~=2.28.2
tenacity~=8.2.2
charset_normalizer==2.1.0
html2text
docx2txt
jsonlines
selenium~=4.9.0
pypdf
cchardet
pillow
tiktoken

